# Conflicts:
#	frontend/lib/src/model/appointment.dart
#	frontend/lib/src/view/components/appointment_component.dart
#	frontend/lib/src/view/components/appointment_list_component.dart
#	frontend/lib/src/view/components/dayview_component.css
#	frontend/lib/src/view/components/dayview_component.dart
#	frontend/lib/src/view/components/dayview_component.html
#	frontend/lib/src/view/services/appointment_service.dart

+++ frontend/lib/src/view/components/appointment_component.dart
@@ -1,3 +1,5 @@
+import 'dart:html';
+
 import 'package:angular/angular.dart';
 import 'package:angular_forms/angular_forms.dart';
 import 'package:angular_router/angular_router.dart';
@@ -7,7 +9,7 @@ import 'package:demo/src/model/appointment.dart';
 import 'package:demo/src/view/routes/route_paths.dart';
 import 'package:demo/src/view/services/appointment_service.dart';
 
-
+import 'package:demo/src/view/components/login_component.dart';
 
 
 @Component(
@@ -19,34 +21,44 @@ import 'package:demo/src/view/services/appointment_service.dart';
 /// Klasse zur verwaltung der Termine
 class AppointmentComponent implements OnActivate {
 
-  AppointmentComponent(this._appointmentService, this._location);
+  AppointmentComponent(this._appointmentService, this._location, this._router);
 
-  Appointment appointment = Appointment(1, "Test Termin", "", "", "", "");
+  Appointment appointment;
   final Location _location;
-
+  bool deleteControl = false;
+  bool submitted = true;
+  final Router _router;
   final AppointmentService _appointmentService;
 
   /// Folgender Code wird immer bei der Aktivierung der Klasse aufgerufen
   @override
   void onActivate(_, RouterState current) async {
-
-
-    final id = getId(current.parameters);
-    if (id != null) {
-      appointment = await (_appointmentService.get(id));
+    if (!LoginComponent.loggedIn) {
+      await _router.navigate('/login');
+    } else {
+      final id = getId(current.parameters);
+      if (id != null) {
+        appointment = await _appointmentService.get(id);
+      }
     }
   }
+
+  void onSubmit() => submitted = true;
 
   /// Methode zum speichern, der änderungen die man im Termin vorgenommen hat
   Future<void> save() async {
     await _appointmentService.update(appointment);
-    goBack();
   }
 
   ///Methode zum löschen von Terminen
   Future<void> delete() async {
-    await _appointmentService.delete(appointment.id);
-    goBack();
+    if (deleteControl==true) {
+      await _appointmentService.delete(appointment.id);
+      goBack();
+    } else {
+      deleteControl = true;
+
+    }
   }
 
   ///Methode, die die übergeordnete ansicht anzeigt
